Major bugfix on display + minor bugfix for uniform export
getBiotype was not returning anything, resulting in data not being
exported, and prevented any display.
Uniform is not exported when it is created, but when other datas are
exported. It prevents the loading of the page when it is not necessary.
Will be useful when several NaviCom instancies are used together in a
NaviGroup and only comparisons are wanted.