Configuration of display for every dataset on export
The display configuration can be choosen with the DisplayConfig class,
and passed as an argument to the NaviCom class. Display settings are
configured in NaviCell right after data are exported.
Export text has been moved to occur before export instead of after.

Fix bug in signif for negative values
Fix typo for continuous copy number