displayMutations + easy selection of biotype
Add displayMutations function.
Add selectDataFromBiotype to select method names from common biological
types names.
Fix a bug in color configuration that when the min/max value is zero the
list created was incorrect.
Fix a bug in glyph size configuration that put a value of zero to the
first slot even when it was not an NA.