Add completeDisplay function + mutations handling patches
Add the completeDisplay function to display as many data as possible on
one map.
Had to patch the mutations datatables management to be able to display
mutations data. Will need a more stable fix.
Fix a bug for all nan tables and text tables in _configDisplay.
Fix getGenomicData typo that created an infinite recursion.
Add getmiRNAData