Change number of significant digits kept for annotations ranges
Some dataset have years as annotations, and if too many years are
present it needs to be reduced. And 4 digits are necessary to
discriminates between years, so the former 3 digits significance was not
enough.
Be more flexible on the maximum number of groups, only regroup for
annotations with really a lot of values.