Define specific glyph configurations
Apply specific configuration for data that are to be use with glyph.
When a data is used as a glyph, this configuration sets one color and
one shape for all values, leaving only the size to see the data.
miRNA, mutations, proteomics and methylation have by default those glyph
configurations, and will only be used with glyphs