Improve size settings for predetermined glyphs
Make sure value is never nan, making the percentile calculation always
correct.
Change the rule for the division to replace the 0, so that the intervals
are more meaningful. Keep the first interval at 1/nb_of_samples for 1
time mutated samples still appear.
Don't apply min_size for glyphs if there are negative values.